+++ src/ingest/subscriptions/post.js
@@ -24,7 +24,7 @@ module.exports = async (req, res) => {
 	try {
 		// generate subscription name
 		const subIdent = 'subscription'
-		const prefix = `${config.pubsubPrefix}${subIdent}.${config.stage}`
+		const prefix = `${config.pubsubPrefix}${subIdent}.`
 		const topicName = req.body.topic
 
 		// check existence of user institution