+++ CHANGELOG.md
@@ -2,12 +2,26 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.co,m/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.0.0-beta] - 2021-03-24
+
+🚧 BREAKING CHANGES for `serviceIds` 🚧
+
+### Added
+
+- New detailed output when posting events for each service (`published`, `blocked`, `failed`)
+
+### Changed
+
+- No longer using `serviceIds` as required keys
+- Now using `services` with required fields `type`, `externalId` and `publisherId` to identify a channel
+- Improved log output in JSON format for better monitoring
 
 ## [0.1.7] - 2021-03-18
 
-### Changes
+### Changed
 
 - Add auth verification for unit tests
 - Check `content-type` in unit tests
@@ -15,14 +29,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.6] - 2021-03-17
 
-### Changes
+### Changed
 
 - Allow optional fields in POST /events to be `null`
 - Remove field `isInternal` from POST /events
 
 ## [0.1.5] - 2021-03-17
 
-### Changes
+### Changed
 
 - Preventing errors when `institution.name` isn't properly set in the user account
 - Enforcing separation between dev/prod topics and subscriptions
@@ -30,7 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.4] - 2021-03-16
 
-### Changes
+### Changed
 
 - Removed `attribution` from required media fields of new events
 - Added OpenAPI documentation to the docs
@@ -39,7 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.3] - 2021-03-16
 
-### Changes
+### Changed
 
 - Hotfix `content-type` bug for error responses
 - Updated endpoint structure for `/events`